+++ zanata-war/src/test/java/org/zanata/service/impl/TranslationUpdatedManagerTest.java
@@ -100,7 +100,7 @@ public class TranslationUpdatedManagerTest {
     private LocaleId localeId = LocaleId.DE;
     private String versionSlug = "versionSlug";
     private String projectSlug = "projectSlug";
-    private int wordCount = 10;
+    private Long wordCount = 10L;
     private ContentState oldState = ContentState.New;
     private ContentState newState = ContentState.Translated;
     private boolean isDisplayUserEmail = true;
@@ -140,7 +140,7 @@ public class TranslationUpdatedManagerTest {
         when(version.getProject()).thenReturn(project);
         when(project.getSlug()).thenReturn(projectSlug);
         when(project.getWebHooks()).thenReturn(webHooks);
-        when(textFlowDAO.getWordCount(tfId)).thenReturn(wordCount);
+        when(textFlowDAO.getWordCount(tfId)).thenReturn(wordCount.intValue());
         when(userService.transferToUser(account, isDisplayUserEmail))
             .thenReturn(user);
         when(textFlowTargetDAO.findById(tftId)).thenReturn(target);
@@ -153,17 +153,17 @@ public class TranslationUpdatedManagerTest {
 
         WordStatistic stats = new WordStatistic(10, 10, 10, 10, 10);
         WordStatistic oldStats = StatisticsUtil.copyWordStatistic(stats);
-        oldStats.decrement(newState, wordCount);
-        oldStats.increment(oldState, wordCount);
+        oldStats.decrement(newState, wordCount.intValue());
+        oldStats.increment(oldState, wordCount.intValue());
 
         when(translationStateCache.getDocumentStatistics(docId, localeId)).
             thenReturn(stats);
-        when(textFlowDAO.getWordCount(tfId)).thenReturn(wordCount);
+        when(textFlowDAO.getWordCount(tfId)).thenReturn(wordCount.intValue());
 
         DocumentLocaleKey key =
             new DocumentLocaleKey(docId, localeId);
 
-        Map<ContentState, Integer> contentStates = new HashMap<>();
+        Map<ContentState, Long> contentStates = new HashMap<>();
         contentStates.put(newState, wordCount);
         contentStates.put(oldState, wordCount);
